sample_app/pyproject.toml
[tool.poetry]
name = "sample_app"
version = "0.1.0"
description = "Sample app for digitial ocean k8s challenge"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiokafka = "^0.7.2"
black = "^21.12b0"
elasticsearch = {extras = ["async"], version = "^7.16.2"}
fastapi = "^0.70.1"
fastavro = "^1.4.8"
pydantic = "^1.8.2"
python-dotenv = "^0.19.2"
uvicorn = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
